Move delimiter and scope guard to common
