+++ ohantsev.vladimir/common/delimiter.cpp
@@ -1,0 +1,18 @@
+#include "delimiter.h"
+#include <iostream>
+
+std::istream& ohantsev::operator>>(std::istream& in, DelimiterIO&& dest)
+{
+  std::istream::sentry sentry(in);
+  if (!sentry)
+  {
+    return in;
+  }
+  char c;
+  in >> c;
+  if (in && (c != dest.exp))
+  {
+    in.setstate(std::ios::failbit);
+  }
+  return in;
+}
