+++ ohantsev.vladimir/common/iofmtguard.cpp
