+++ ohantsev.vladimir/common/iofmtguard.h
